[Map] Now changing input mode in fullscreen mode.
Also release the mouse, and recapture it when exiting the mode.
