--- Snowy-Path/Assets/Scenes/Dev/QB/04-Map/00-Common/MapTool.cs
@@ -1,21 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.InputSystem;
-
-public class MapTool : MonoBehaviour
-{
-    public GameObject mapUI;
-
-    void Start()
-    {
-    }
-
-    void Update()
-    {
-        var keyboard = Keyboard.current;
-        if (keyboard.mKey.wasPressedThisFrame) {
-            mapUI.SetActive(!mapUI.activeSelf);
-        }
-    }
-}
