+++ Snowy-Path/Assets/Scenes/Dev/QB/04-Map/02-Fullscreen.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9f1097befa292fe4e80dbe6595304958
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
