+++ Snowy-Path/Assets/Scenes/Dev/QB/04-Map/02-Fullscreen/MapTool.cs.meta
