+++ Snowy-Path/Assets/Scenes/Dev/QB/04-Map/02-Fullscreen/QBMap_Fullscreen.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ec8e75d048d226f43bae056e26f6209f
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
